refactor: refactor usecase
