+++ usecase/src/model.rs
@@ -1,3 +1,4 @@
 pub mod event;
 pub mod spot;
+pub mod traffic;
 pub mod visitor;
