+++ usecase/src/usecase.rs
@@ -1,3 +1,8 @@
 pub mod admin;
 pub mod error;
+pub mod event;
+pub mod image;
+pub mod palette;
+pub mod spot;
+pub mod traffic;
 pub mod visitor;
