+++ usecase/src/usecase/visitor.rs
@@ -1,14 +1,6 @@
-use std::str::FromStr;
-
 use async_trait::async_trait;
-use futures::future::join_all;
-use rand::rngs::StdRng;
-use rand::seq::SliceRandom;
-use rand::SeedableRng;
 use repaint_server_model::event::Event;
-use repaint_server_model::event_spot::EventSpot;
 use repaint_server_model::id::Id;
-use repaint_server_model::visitor_image::{CurrentImage, Image as VisitorImage};
 use repaint_server_model::AsyncSafe;
 use teloc::inject;
 
@@ -38,64 +30,30 @@ pub trait VisitorUsecase: AsyncSafe {
         &self,
         visitor_identification: VisitorIdentification,
     ) -> Result<(), Error>;
-
-    async fn list_image(
-        &self,
-        visitor_identification: VisitorIdentification,
-    ) -> Result<Vec<Id<VisitorImage>>, Error>;
-
-    async fn get_current_image(
-        &self,
-        visitor_identification: VisitorIdentification,
-    ) -> Result<Id<CurrentImage>, Error>;
-
-    async fn set_current_image(
-        &self,
-        visitor_identification: VisitorIdentification,
-        image_id: Id<VisitorImage>,
-    ) -> Result<(), Error>;
-
-    async fn drop_palette(
-        &self,
-        visitor_identification: VisitorIdentification,
-        spot_id: Id<EventSpot>,
-    ) -> Result<(), Error>;
-
-    async fn pick_palette(
-        &self,
-        visitor_identification: VisitorIdentification,
-    ) -> Result<(), Error>;
 }
 
 #[derive(Debug)]
-pub struct VisitorUsecaseImpl<R> {
+pub struct VisitorUsecaseImpl<R, F> {
     repo: R,
+    firestore: F,
 }
 
 #[inject]
-impl<R> VisitorUsecaseImpl<R>
+impl<R, F> VisitorUsecaseImpl<R, F>
 where
-    R: VisitorRepository
-        + EventRepository
-        + SpotRepository
-        + ImageRepository
-        + PaletteRepository
-        + Firestore,
+    R: VisitorRepository + EventRepository + SpotRepository + ImageRepository + PaletteRepository,
+    F: Firestore,
 {
-    pub fn new(repo: R) -> Self {
-        Self { repo }
+    pub fn new(repo: R, firestore: F) -> Self {
+        Self { repo, firestore }
     }
 }
 
 #[async_trait]
-impl<R> VisitorUsecase for VisitorUsecaseImpl<R>
+impl<R, F> VisitorUsecase for VisitorUsecaseImpl<R, F>
 where
-    R: VisitorRepository
-        + EventRepository
-        + SpotRepository
-        + ImageRepository
-        + PaletteRepository
-        + Firestore,
+    R: VisitorRepository + EventRepository + SpotRepository + ImageRepository + PaletteRepository,
+    F: Firestore,
 {
     async fn join_event(
         &self,
@@ -119,7 +77,9 @@ where
             VisitorRepository::create(&self.repo, event.id, registration_id.clone()).await?;
         let palettes = PaletteRepository::get(&self.repo, visitor.id).await?;
 
-        Firestore::subscribe_register_log(&self.repo, event_id, visitor.visitor_id).await?;
+        self.firestore
+            .subscribe_register_log(event_id, visitor.visitor_id)
+            .await?;
 
         Ok((
             EventResponse {
@@ -169,12 +129,12 @@ where
         let image_id = ImageRepository::get_visitor_image(&self.repo, visitor.id).await?;
         let current_image_id = ImageRepository::get_current_image(&self.repo, visitor.id).await?;
 
-        Firestore::subscribe_initialize_log(
-            &self.repo,
-            visitor_identification.event_id,
-            visitor_identification.visitor_id,
-        )
-        .await?;
+        self.firestore
+            .subscribe_initialize_log(
+                visitor_identification.event_id,
+                visitor_identification.visitor_id,
+            )
+            .await?;
 
         Ok((
             EventResponse {
@@ -211,216 +171,6 @@ where
                     message: format!("{} is invalid id", visitor_identification.visitor_id),
                 })?;
         let _ = VisitorRepository::delete(&self.repo, visitor.id).await?;
-
-        Ok(())
-    }
-
-    async fn list_image(
-        &self,
-        visitor_identification: VisitorIdentification,
-    ) -> Result<Vec<Id<VisitorImage>>, Error> {
-        let event = EventRepository::get(&self.repo, visitor_identification.event_id)
-            .await?
-            .ok_or(Error::BadRequest {
-                message: format!("{} is invalid id", visitor_identification.event_id),
-            })?;
-        let visitor =
-            VisitorRepository::get(&self.repo, event.id, visitor_identification.visitor_id)
-                .await?
-                .ok_or(Error::BadRequest {
-                    message: format!("{} is invalid id", visitor_identification.visitor_id),
-                })?;
-        let default = ImageRepository::list_default_image(&self.repo, event.id).await?;
-        let visitor = ImageRepository::get_visitor_image(&self.repo, visitor.id).await?;
-        let mut images = default
-            .iter()
-            .filter_map(|&i| Id::<VisitorImage>::from_str(&i.to_string()).ok())
-            .collect::<Vec<_>>();
-        if let Some(visitor) = visitor {
-            images.push(visitor);
-        };
-
-        Ok(images)
-    }
-
-    async fn get_current_image(
-        &self,
-        visitor_identification: VisitorIdentification,
-    ) -> Result<Id<CurrentImage>, Error> {
-        let event = EventRepository::get(&self.repo, visitor_identification.event_id)
-            .await?
-            .ok_or(Error::BadRequest {
-                message: format!("{} is invalid id", visitor_identification.event_id),
-            })?;
-        let visitor =
-            VisitorRepository::get(&self.repo, event.id, visitor_identification.visitor_id)
-                .await?
-                .ok_or(Error::BadRequest {
-                    message: format!("{} is invalid id", visitor_identification.visitor_id),
-                })?;
-        let current_image_id = ImageRepository::get_current_image(&self.repo, visitor.id).await?;
-
-        Ok(current_image_id)
-    }
-
-    async fn set_current_image(
-        &self,
-        visitor_identification: VisitorIdentification,
-        image_id: Id<VisitorImage>,
-    ) -> Result<(), Error> {
-        let event = EventRepository::get(&self.repo, visitor_identification.event_id)
-            .await?
-            .ok_or(Error::BadRequest {
-                message: format!("{} is invalid id", visitor_identification.event_id),
-            })?;
-        let visitor =
-            VisitorRepository::get(&self.repo, event.id, visitor_identification.visitor_id)
-                .await?
-                .ok_or(Error::BadRequest {
-                    message: format!("{} is invalid id", visitor_identification.visitor_id),
-                })?;
-        let _ = ImageRepository::set_current_image(&self.repo, visitor.id, image_id).await?;
-
-        Ok(())
-    }
-
-    async fn drop_palette(
-        &self,
-        visitor_identification: VisitorIdentification,
-        spot_id: Id<EventSpot>,
-    ) -> Result<(), Error> {
-        let event = EventRepository::get(&self.repo, visitor_identification.event_id)
-            .await?
-            .ok_or(Error::BadRequest {
-                message: format!("{} is invalid id", visitor_identification.event_id),
-            })?;
-        let visitor =
-            VisitorRepository::get(&self.repo, event.id, visitor_identification.visitor_id)
-                .await?
-                .ok_or(Error::BadRequest {
-                    message: format!("{} is invalid id", visitor_identification.visitor_id),
-                })?;
-        let palettes = PaletteRepository::get(&self.repo, visitor.id).await?;
-        let took_photo = ImageRepository::get_visitor_image(&self.repo, visitor.id)
-            .await?
-            .is_some();
-        let is_bonus = SpotRepository::get_bonus_state(&self.repo, event.id, spot_id).await?;
-
-        Firestore::subscribe_visitor_log(
-            &self.repo,
-            visitor_identification.event_id,
-            visitor_identification.visitor_id,
-            spot_id,
-            palettes.len(),
-            took_photo,
-        )
-        .await?;
-        Firestore::subscribe_visitor(
-            &self.repo,
-            visitor_identification.event_id,
-            visitor_identification.visitor_id,
-            spot_id,
-        )
-        .await?;
-
-        let mut rng = {
-            let rng = rand::thread_rng();
-            StdRng::from_rng(rng).unwrap()
-        };
-
-        if is_bonus {
-            let palettes =
-                Firestore::get_palettes(&self.repo, visitor_identification.event_id, spot_id)
-                    .await?;
-            let palettes = palettes
-                .choose_multiple(&mut rng, 2)
-                .cloned()
-                .collect::<Vec<_>>();
-            let p = palettes
-                .iter()
-                .map(|&p| PaletteRepository::set(&self.repo, visitor.id, p));
-            join_all(p)
-                .await
-                .into_iter()
-                .collect::<Result<Vec<_>, _>>()?;
-
-            let palettes = palettes.choose_multiple(&mut rng, 2).cloned().collect();
-            Firestore::subscribe_palettes(
-                &self.repo,
-                visitor_identification.event_id,
-                spot_id,
-                palettes,
-            )
-            .await?;
-        } else {
-            let palette =
-                Firestore::get_palette(&self.repo, visitor_identification.event_id, spot_id)
-                    .await?;
-            let _ = PaletteRepository::set(&self.repo, visitor.id, palette).await?;
-
-            let palette = palettes.choose(&mut rng).cloned().unwrap();
-            Firestore::subscribe_palette(
-                &self.repo,
-                visitor_identification.event_id,
-                spot_id,
-                palette,
-            )
-            .await?;
-        }
-
-        Ok(())
-    }
-
-    async fn pick_palette(
-        &self,
-        visitor_identification: VisitorIdentification,
-    ) -> Result<(), Error> {
-        let event = EventRepository::get(&self.repo, visitor_identification.event_id)
-            .await?
-            .ok_or(Error::BadRequest {
-                message: format!("{} is invalid id", visitor_identification.event_id),
-            })?;
-        let visitors = VisitorRepository::list(&self.repo, event.id).await?;
-
-        let p = visitors
-            .iter()
-            .map(|v| PaletteRepository::get(&self.repo, v.id));
-        let palettes = join_all(p)
-            .await
-            .into_iter()
-            .collect::<Result<Vec<_>, _>>()?
-            .into_iter()
-            .flatten()
-            .collect::<Vec<_>>();
-
-        let event = EventRepository::get(&self.repo, visitor_identification.event_id)
-            .await?
-            .ok_or(Error::BadRequest {
-                message: format!("{} is invalid id", visitor_identification.event_id),
-            })?;
-        let visitor =
-            VisitorRepository::get(&self.repo, event.id, visitor_identification.visitor_id)
-                .await?
-                .ok_or(Error::BadRequest {
-                    message: format!("{} is invalid id", visitor_identification.visitor_id),
-                })?;
-
-        let visitor_palettes = PaletteRepository::get(&self.repo, visitor.id)
-            .await?
-            .into_iter()
-            .collect::<Vec<_>>();
-
-        let mut rng = {
-            let rng = rand::thread_rng();
-            StdRng::from_rng(rng).unwrap()
-        };
-
-        while let Some(&palette) = palettes.choose(&mut rng) {
-            if !visitor_palettes.contains(&palette) {
-                let _ = PaletteRepository::set(&self.repo, visitor.id, palette);
-                break;
-            }
-        }
 
         Ok(())
     }
